jatko2

+++ app/src/main/java/com/example/carbgg/InsulinCalculator.java
@@ -7,16 +7,16 @@ public class InsulinCalculator extends Settings{
     private float insulinAmount;
 
 
-    public InsulinCalculator(int carbs)
+    public InsulinCalculator(float carbs)
     {
-        this.carbs = (float)carbs;
+        this.carbs = carbs;
         this.insulinEfficiency = super.recallValues();
     }
 
     public int calculateInsulin(){
         this.carbsDivision = carbs / 10;
         this.insulinAmount = this.carbsDivision / insulinEfficiency;
-        return  (int)insulinAmount;
+        return (int)insulinAmount;
     }
 
 }
